Add BDD testing framework with Behave for product API
- Add Dockerfile.behave and docker-compose.behave.yml for test environment
- Implement comprehensive product feature tests with Gherkin scenarios
- Add product_steps.py with step definitions for CRUD operations
- Include health check endpoint and test data cleanup endpoint
- Update dependencies to include behave testing framework
- Add Taskfile task for running BDD tests

+++ api/models.py
@@ -48,5 +48,5 @@ class ProductResponse(BaseModel):
     product_id: str
     name: str
     description: str | None = None
-    price: int
+    price: Decimal
     created_at: datetime
